projects/betting-bot-telegram/requirements.txt
aiogram==3.13.1
python-dotenv==1.0.1
asyncpg==0.29.0
sqlalchemy==2.0.35
apscheduler==3.10.4
google-auth==2.35.0
google-auth-oauthlib==1.2.1
google-api-python-client==2.149.0
gspread==6.1.2
yadisk==3.1.0
aiohttp==3.10.8
python-dateutil==2.9.0
pytz==2024.1
psycopg2-binary==2.9.9
